# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BugShot API Documentation",
        "description": "OpenApi documentation using swagger for BugShot project.",
        "contact": {
            "email": "info@it-michel.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.test/api/v1"
        }
    ],
    "paths": {
...
        "/bugs/{bug_id}/screenshots": {
            "get": {
                "tags": [
                    "Screenshot"
                ],
                "summary": "All screenshots of the bug.",
                "operationId": "allScreenshots",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Screenshot"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Screenshot"
                ],
                "summary": "Store one screenshots.",
                "operationId": "storeScreenshot",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "base64"
                                ],
                                "properties": {
                                    "position_x": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "position_y": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "web_position_x": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "web_position_y": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "base64": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screenshot"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bugs/{bug_id}/screenshots/{screenshot_id}": {
            "get": {
                "tags": [
                    "Screenshot"
                ],
                "summary": "Show one screenshots.",
                "operationId": "showScreenshot",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    },
                    {
                        "name": "screenshot_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Screenshot/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screenshot"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Screenshot"
                ],
                "summary": "Update one screenshots.",
                "operationId": "updateScreenshot",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    },
                    {
                        "name": "screenshot_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Screenshot/properties/id"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "position_x",
                                    "position_y",
                                    "base64"
                                ],
                                "properties": {
                                    "position_x": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "position_y": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "web_position_x": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "web_position_y": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "base64": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screenshot"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Screenshot"
                ],
                "summary": "Delete one screenshots.",
                "operationId": "deleteScreenshot",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    },
                    {
                        "name": "screenshot_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Screenshot/properties/id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Screenshot": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "bug_id": {
                        "description": "The id of the but to which the object belongs.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "designation": {
                        "description": "The screenshot name.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "url": {
                        "description": "The lacal path where the file is stored. ",
                        "type": "string"
                    },
                    "position_x": {
                        "description": "The x coordinate value for the screenshot marker.",
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "position_y": {
                        "description": "The y coordinate value for the screenshot marker.",
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "web_position_x": {
                        "description": "The x coordinate value of the marker relative to web page top.",
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "web_position_y": {
                        "description": "The y coordinate value of the marker relative to web page top.",
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "The creation date.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The last date when the resource was changed.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "The deletion date.",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}